more warnings fixed for vs2010

+++ include/us_http_server.h
@@ -66,8 +66,8 @@ extern "C" {
         us_buffer_t *m_response_content;
 
         us_http_server_handler_state_t m_state;
-        int32_t m_byte_count;
-        int32_t m_todo_count;
+        size_t m_byte_count;
+        size_t m_todo_count;
 
         struct us_webapp_director_s *m_director;
